pyproject.toml
[tool.poetry]
name = "charter-compiler"
version = "0.1.0"
description = "Agent compiler with orchestration-level memoization for LLM inference"
authors = ["Your Name"]

[tool.poetry.dependencies]
python = "^3.10"
libcst = "^1.0.0"
networkx = "^3.0"
scalpel = "^0.2.0"
numpy = "^1.24.0"
tritonclient = {extras = ["grpc", "http"], version = "^2.40.0"}
prometheus-client = "^0.19.0"
pyyaml = "^6.0"
click = "^8.1.0"
rich = "^13.0.0"
pydantic = "^2.0.0"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true


